+++ python/python/bystro/ancestry/tests/test_adversarial_autoencoder.py
@@ -1,38 +1,32 @@
 import numpy as np
-
+import scipy.stats as st
 from bystro.ancestry.adversarial_autoencoder import AdversarialAutoencoder
 
+def generate_data():
+    rng = np.random.default_rng(2021)
+    N = 10000
+    p = 100
+    L = 3
+    sigma = 1.0
+    W_base = st.ortho_group.rvs(p)
+    W = W_base[:L]
+    lamb = rng.gamma(1, 1, size=L) + 1
+    for ll in range(L):
+        W[ll] = W[ll] * lamb[ll]
+    S_train = rng.normal(size=(N, L))
+    X_hat = np.dot(S_train, W)
+    X_noise = sigma * rng.normal(size=(N, p))
+    X_train = X_hat + X_noise
+    return W, sigma, X_train, S_train
 
-def test_adversarial_autoencoder_convergence():
-    # Parameters
-    n_samples = 1000
-    observation_dimension = 10
-    n_components = 3
-    n_iterations = 100
-    batch_size = 50
+def test_adversarial_autoencoder():
+    W,sigma, X,S_train = generate_data()
+    model = AdversarialAutoencoder(3)
+    model.fit(X)
 
-    # Generate synthetic data (simple Gaussian distribution for testing)
-    rng = np.random.default_rng()  # Create a random number generator instance
-    X_synthetic = rng.standard_normal((n_samples, observation_dimension)).astype(np.float32)
+    S_est = model.transform(X.astype(np.float32))
+    plt.scatter(S_est[:,0],S_est[:,1])
 
-    # Instantiate the autoencoder
-    AAE = AdversarialAutoencoder(n_components)
-    AAE.training_options["n_iterations"] = n_iterations
-    AAE.training_options["batch_size"] = batch_size
-    AAE.training_options["learning_rate"] = 0.001
-
-    # Train the autoencoder
-    AAE.fit(X_synthetic)
-
-    # Assert that the generative and discriminative losses have decreased
-    assert AAE.losses_generative[-1] < AAE.losses_generative[0], "Generative loss did not decrease"
-    assert (
-        AAE.losses_discriminative[-1] < AAE.losses_discriminative[0]
-    ), "Discriminative loss did not decrease"
-
-    assert AAE.encoder is not None, "Encoder was not initialized"
-    assert AAE.decoder is not None, "Decoder was not initialized"
-
-
-# Run the test
-test_adversarial_autoencoder_convergence()
+	
+if __name__ == "__main__":
+    test_adversarial_autoencoder()
